pyproject.toml
[project]
name = "atdata"
version = "0.3.0b1"
description = "A loose federation of distributed, typed datasets"
readme = "README.md"
authors = [
    { name = "Maxine Levesque", email = "hello@maxine.science" },
    { name = "Maxine @ Forecast Bio", email = "maxine@forecast.bio" }
]
requires-python = ">=3.12"
dependencies = [
    "atproto>=0.0.65",
    "boto3>=1.41.5",
    "fastparquet>=2024.11.0",
    "libipld>=3.3.2",
    "msgpack>=1.1.2",
    "numpy>=2.3.4",
    "ormsgpack>=1.11.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis-om>=0.3.5",
    "requests>=2.32.5",
    "s3fs>=2025.12.0",
    "schemamodels>=0.9.1",
    "tqdm>=4.67.1",
    "typer>=0.21.1",
    "webdataset>=1.0.2",
]

[project.optional-dependencies]
atmosphere = [
    "atproto>=0.0.55",
]
postgres = [
    "psycopg[binary]>=3.1",
]

[project.scripts]
atdata = "atdata:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = "--cov=atdata --cov-report=html"
markers = [
    "network: marks tests as requiring network access (deselect with '-m \"not network\"')",
    "slow: marks tests as slow running",
    "benchmark: performance benchmarks",
    "bench_serial: serialization benchmarks (μs scale)",
    "bench_index: index provider benchmarks (μs–ms scale)",
    "bench_io: shard read/write benchmarks (ms scale)",
    "bench_query: query system benchmarks (ms scale)",
    "bench_s3: S3 storage benchmarks via moto (ms+ scale)",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "moto[s3]>=5.0.29",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "quartodoc>=0.11.1",
    "pytest-benchmark>=5.1.0",
    "ruff>=0.14.13",
]
